Fix QCOMPARE() mixed-type comparisons to conform with the strict pre-Qt5.10 API
See https://stackoverflow.com/q/29102299
    https://codereview.qt-project.org/c/qt/qtbase/+/192269/

+++ src/test/targetreportextractortest/targetreportextractortest.cpp
@@ -499,8 +499,8 @@ void TargetReportExtractorTest::test()
 
     QVERIFY(trqueue.size() == tgtRepsOut.size());
 
-    QCOMPARE(counter.in_, recsIn.size());
-    QCOMPARE(counter.out_, tgtRepsOut.size());
+    QCOMPARE(counter.in_, static_cast<quint32>(recsIn.size()));
+    QCOMPARE(counter.out_, static_cast<quint32>(tgtRepsOut.size()));
 
     for (int i = 0; i < tgtRepsOut.size(); ++i)
     {
